+++ examples/data-sources/iosxe_flow_exporter/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxe_flow_exporter" "example" {
+  name = "EXPORTER1"
+}