+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add support for descriptions to `iosxe_prefix_list` resource and data source
 - Add `advertisement_interval` attribute to `iosxe_bgp_ipv4_unicast_vrf_neighbor` resource and data source
 - Add `iosxe_flow_record` resource and data source
+- Add `iosxe_flow_exporter` resource and data source
 
 ## 0.5.5